fix(migrations): all prices are decimals

+++ database/migrations/2025_12_14_222745_create_movements_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->foreignId('guest_id')->constrained();
             $table->foreignId('payment_id')->nullable()->constrained();
             $table->foreignId('article_id')->nullable()->constrained();
-            $table->integer('amount');
+            $table->decimal('amount', 8, 2);
             $table->string('type');
             $table->json('meta')->nullable();
         });
