+++ database/migrations/2026_02_02_194300_create_payments_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->string('stripe_status')->nullable();
             $table->json('stripe_data');
             $table->string('purpose')->nullable();
-            $table->integer('amount');
+            $table->decimal('amount', 8, 2);
         });
     }
